download/pyproject.toml
[tool.poetry]
name = "stopreasons"
version = "0.3.0"
description = "Module to fine tune a NLP model that assigns one or multiple classes of why a clinical trial has stopped early. The base architecture is a BERT model fine tuned on a dataset curated by the Open Targets team."
authors = ["Irene López <irene.lopezs@protonmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
datasets = "^2.4.0"
transformers = {extras = ["tf-cpu"], version = "^4.21.3"}
spark-nlp = "4.1.0"
pyspark = "3.3.0"
pathlib = "^1.0.1"
ipython = "^8.7.0"
scikit-learn = "^1.2.0"
torch = "1.12.1"
tensorflow = "2.10"
huggingface-hub = "^0.12.0"
typer = "^0.7.0"
evaluate = "^0.4.0"
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.3.0"
radian = "^0.6.5"
omegaconf = "^2.3.0"
statsmodels = "^0.14.0"
matplotlib = "^3.7.1"
plotly-express = "^0.4.1"
nbformat = "^5.9.0"
chart-studio = "^1.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pydocstyle = "^6.3.0"
mypy = "^1.3.0"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
